+++ app/Http/Controllers/UserController.php
@@ -10,7 +10,7 @@ class UserController extends Controller
     public function getUserinfo($id)
     {
         $usInfo = User::where('id',$id)->first();
-        return view('webclient.userinfo',['Usinfo'=>$usInfo]);
+        return view('webclient.Index.Updateinfo',['Usinfo'=>$usInfo]);
     }
     public function getUser()
     {
@@ -21,17 +21,30 @@ class UserController extends Controller
     {
 
         $user = User::where('id',$id)->first();
+        $user->user_fname = $request->user_fname;
+        $user->user_email = $request->user_email;
         $user->user_gender = $request->user_gender;
         $user->user_address = $request->user_address;
         $user->user_phone = $request->user_phone;
-        $user->user_id_card = $request->user_id_card;
+        if($request->password != "")
+        {
+            $user->password = bcrypt($request->password);
+        }
         $user->save();
-        return redirect()->back()->with('Thongbao','Bạn đã xóa thành công !');
+        return redirect()->back()->with('Thongbao','Bạn đã cập nhập thông tin thành công !');
     }
     public function getDeleteUser($id)
     {
-        $user = User::where('id',$id);
-        $user->delete();
-        return redirect('/admin/dashboard/membermanager')->with('Thongbao','Bạn đã xóa thành công !');
+        $user = User::where('id', $id)->first();
+        if(Auth::user()->id != $user->id) {
+            if (Auth::user()->role_id > $user->role_id) {
+                $user->delete();
+                return redirect('/admin/dashboard/membermanager')->with('Thongbao', 'Bạn đã xóa thành công !');
+            } else
+                return redirect('/admin/dashboard/membermanager')->with('Loi', 'Bạn chỉ có thể xóa người có quyền nhỏ hơn bạn !');
+        }
+        else
+            return redirect('/admin/dashboard/membermanager')->with('Loi', 'Bạn không thể xóa bản thân !');
     }
+
 }
